requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.1
django-crispy-forms==1.9.2
mysqlclient==2.0.1
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.2
Pillow==7.2.0
plotly==4.11.0
pycodestyle==2.6.0
PyMySQL==0.10.1
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Wand==0.6.3
xlwt==1.3.0
